DEsign Pattern

+++ src/main/java/com/parcel/parcelcosting/service/ParcelService.java
@@ -2,6 +2,8 @@ package com.parcel.parcelcosting.service;
 
 import com.parcel.parcelcosting.entity.Parcel;
 
+import java.math.BigDecimal;
+
 public interface ParcelService {
-    public Double getCost(Parcel parcel);
+    public BigDecimal getCost(Parcel parcel);
 }
